// webapp/src/app/data/resource/model/resource-step.model.ts
// Mockup: https://sketch.cloud/s/3lW3m/a/Zy5PD5 
// Model for the "Step-by-Step"
export interface ResourceStepModel {
    // UNKNOWN, New concept in DL 2.0?
    stepNumber: number;

    // UNKNOWN, New concept in DL 2.0?
    title: string;

    // UNKNOWN, New concept in DL 2.0?
    // as html?
    content: string;
}